feat(ts-sdk): extract error details
