+++ ts-sdk/src/utils/index.ts
@@ -1,0 +1,1 @@
+export { extractErrorDetails } from './extract-error-details.js'
